chore: move book components
The book components should be within their own folder, which is necessary now that we have more components.
